Refactor clowder yaml validation functions
